move generated to api
