+++ buf.gen.yaml
@@ -1,10 +1,10 @@
 version: v1
 plugins:
   - name: go
-    out: generated
+    out: .
     opt: paths=source_relative
   - remote: buf.build/grpc/plugins/go:v1.2.0-1
-    out: generated
+    out: .
     opt:
       - paths=source_relative
       - require_unimplemented_servers=false
