+++ buf.lock
@@ -1,0 +1,7 @@
+# Generated by buf. DO NOT EDIT.
+version: v1
+deps:
+  - remote: buf.build
+    owner: googleapis
+    repository: googleapis
+    commit: faacf837d7304c58b7c9020c7807fa6e
