+++ internal/resthandler/session.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/go-seidon/chariot/generated/restapp"
+	"github.com/go-seidon/chariot/api/restapp"
 	"github.com/go-seidon/chariot/internal/session"
 	"github.com/go-seidon/provider/status"
 	"github.com/labstack/echo/v4"
