+++ internal/restmiddleware/basicauth.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/go-seidon/chariot/generated/restapp"
+	"github.com/go-seidon/chariot/api/restapp"
 	"github.com/go-seidon/chariot/internal/auth"
 	"github.com/go-seidon/provider/serialization"
 	"github.com/go-seidon/provider/status"
