+++ internal/restmiddleware/basicauth_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/go-seidon/chariot/generated/restapp"
+	"github.com/go-seidon/chariot/api/restapp"
 	"github.com/go-seidon/chariot/internal/auth"
 	mock_auth "github.com/go-seidon/chariot/internal/auth/mock"
 	"github.com/go-seidon/chariot/internal/restmiddleware"
