+++ internal/restmiddleware/sessionauth_test.go
@@ -5,7 +5,7 @@ import (
 	"mime/multipart"
 	"net/http"
 
-	"github.com/go-seidon/chariot/generated/restapp"
+	"github.com/go-seidon/chariot/api/restapp"
 	"github.com/go-seidon/chariot/internal/restmiddleware"
 	"github.com/go-seidon/chariot/internal/session"
 	mock_session "github.com/go-seidon/chariot/internal/session/mock"
