+++ sonar-project.properties
@@ -11,9 +11,6 @@ sonar.tests=.
 sonar.test.inclusions=\
   **/*_test.go, \
   api/**, \
-  generated/**, \
-  internal/queueing/**, \
-  internal/rabbitmq/**, \
   internal/app/config.go
 
 sonar.go.coverage.reportPaths=coverage.out
